feat: integrate ttt_max processing step into the execution pipeline
